fixed weird issue
