+++ app/api/posts/get_posts/route.ts
@@ -23,7 +23,6 @@ export async function GET(req: NextRequest) {
             },
           },
           allImages: true,
-          likes: true,
         },
       })
