+++ app/blog/components/ui/PostList.tsx
@@ -1,11 +1,5 @@
 "use client"
 
-import { buttonVariants } from "@/components/ui/button"
-import {
-  HoverCard,
-  HoverCardContent,
-  HoverCardTrigger,
-} from "@/components/ui/hover-card"
 import FlipNumber from "@/layouts/FlipNumber"
 import Paragraph from "@/layouts/Paragraph"
 import SubTitle from "@/layouts/SubTitle"
@@ -13,10 +7,12 @@ import Title from "@/layouts/Title"
 import Fetcher from "@/lib/fetcher"
 import { BlogPostDT } from "@/lib/types"
 import { motion } from "framer-motion"
-import { SmilePlus, SquareChevronLeft, SquareChevronRight } from "lucide-react"
+import { SquareChevronLeft, SquareChevronRight } from "lucide-react"
 import Link from "next/link"
 import React, { useState } from "react"
 import useSWR, { mutate } from "swr"
+import Emojis from "./Emojis"
+import SkeletonPostList from "./SkeletonPostList"
 
 const PostList = () => {
   const [displayedPosts, setDisplayedPosts] = useState(3)
@@ -219,237 +215,3 @@ const PostList = () => {
 }
 
 export default PostList
-
-// const Emojis = ({ blogID }: { blogID: number }) => {
-//   return (
-//     <HoverCard>
-//       <div className="flex items-center gap-5">
-//         <HoverCardTrigger>
-//           <SmilePlus className="cursor-pointer" />
-//         </HoverCardTrigger>
-//         <div className="flex items-center gap-1">
-//           <div className="bg-hoverDark flex items-center gap-[2px] p-1 rounded-md">
-//             <img src="./heart.png" alt="heart" />
-//             <FlipNumber>{0}</FlipNumber>
-//           </div>
-//           <div className="bg-hoverDark flex items-center gap-[2px] p-1 rounded-md">
-//             <img src="./clap.png" alt="clap" />
-//             <FlipNumber>{0}</FlipNumber>
-//           </div>
-//         </div>
-//         <HoverCardContent
-//           style={{
-//             backgroundColor: "#222222",
-//             borderColor: "transparent",
-//             width: "fit-content",
-//             height: "fit-content",
-//             position: "absolute",
-//             top: "-80px",
-//             padding: 5,
-//           }}
-//         >
-//           <div className="flex items-center gap-2">
-//             <div className="h-[30px] w-[50px] cursor-pointer">
-//               <img
-//                 className="h-full w-full object-contain"
-//                 src="./heart.png"
-//                 alt="heart"
-//               />
-//             </div>
-//             <div className="h-full w-full cursor-pointer">
-//               <img
-//                 className="h-[30px] w-[50px] object-contain"
-//                 src="./clap.png"
-//                 alt="clap"
-//               />
-//             </div>
-//           </div>
-//         </HoverCardContent>
-//       </div>
-//     </HoverCard>
-//   )
-// }
-
-const Emojis = ({ blogID }: { blogID: number }) => {
-  const { data: emojis, mutate } = useSWR(
-    `/api/posts/emojis/${blogID}`,
-    Fetcher
-  )
-  const handleEmojiClick = async (type: string) => {
-    await fetch(`/api/posts/emojis/${blogID}`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ type }),
-    })
-    mutate()
-  }
-
-  return (
-    <HoverCard>
-      <div className="flex items-center gap-5">
-        <HoverCardTrigger>
-          <SmilePlus className="cursor-pointer" />
-        </HoverCardTrigger>
-        <div className="flex items-center gap-1">
-          {emojis?.map((emoji) => (
-            <div
-              key={emoji.emojiID}
-              className="bg-hoverDark flex items-center gap-[2px] p-1 rounded-md"
-            >
-              <img src={`./${emoji.type}.png`} alt={emoji.type} />
-              <FlipNumber>{emoji.count}</FlipNumber>
-            </div>
-          ))}
-        </div>
-        <HoverCardContent
-          style={{
-            backgroundColor: "#222222",
-            borderColor: "transparent",
-            width: "fit-content",
-            height: "fit-content",
-            position: "absolute",
-            top: "-80px",
-            padding: 5,
-          }}
-        >
-          <div className="flex items-center gap-2">
-            <div
-              className="h-[30px] w-[50px] cursor-pointer"
-              onClick={() => handleEmojiClick("heart")}
-            >
-              <img
-                className="h-full w-full object-contain"
-                src="./heart.png"
-                alt="heart"
-              />
-            </div>
-            <div
-              className="h-full w-full cursor-pointer"
-              onClick={() => handleEmojiClick("clap")}
-            >
-              <img
-                className="h-[30px] w-[50px] object-contain"
-                src="./clap.png"
-                alt="clap"
-              />
-            </div>
-          </div>
-        </HoverCardContent>
-      </div>
-    </HoverCard>
-  )
-}
-
-const SkeletonPostList = () => {
-  const skeletonPosts = Array.from({ length: 3 }, (_, index) => index)
-
-  return (
-    <motion.div
-      initial={{ opacity: 0, y: 50 }}
-      animate={{ opacity: 1, y: 0 }}
-      transition={{ duration: 0.5, delay: 0.3 }}
-      className="grid grid-cols-[1fr, 1fr, 1fr, 1fr] grid-rows-[1fr, 1fr, 1fr, 1fr] gap-y-[10px] gap-x-7"
-    >
-      {skeletonPosts.map((index) => (
-        <div
-          key={index}
-          className={`relative bg-gradient-to-r from-neutral-400 to-slate-900 dark:from-[#282828] dark:to-[#000000]  rounded-xl text-white animate-in 
-          ${
-            index === 0 &&
-            "col-start-1 col-end-3 row-start-1 row-end-3 h-[420px] "
-          }
-          ${index === 1 && "col-start-3 col-end-4 h-[200px]"}
-          ${index === 2 && "col-start-3 col-end-4 h-[200px]"}
-        `}
-          style={{ "--index": 3 } as React.CSSProperties}
-        >
-          {/* top */}
-          {index === 0 && (
-            <div
-              className="h-[200px] object-cover rounded-tl-xl rounded-tr-xl"
-              style={{
-                background: "gray",
-              }}
-            ></div>
-          )}
-          {/* top */}
-
-          {/* bottom */}
-          <div
-            className={`flex flex-col p-[10px] w-full ${
-              index === 0 && "h-[308px]"
-            }`}
-          >
-            <div className="flex flex-col gap-[10px]">
-              <div
-                className="font-bold text-lg text-white"
-                style={{
-                  width: "80%",
-                  height: "20px",
-                  background: "gray",
-                }}
-              ></div>
-              <div
-                className="text-sm text-white"
-                style={{
-                  width: "100%",
-                  height: "50px",
-                  background: "gray",
-                }}
-              ></div>
-              <div className="flex flex-col gap-[6px]">
-                <div className="flex items-center justify-between">
-                  <div
-                    className="text-base text-white"
-                    style={{
-                      width: "40%",
-                      height: "20px",
-                      background: "gray",
-                    }}
-                  ></div>
-                  <div
-                    className="text-base text-white"
-                    style={{
-                      width: "40%",
-                      height: "20px",
-                      background: "gray",
-                    }}
-                  ></div>
-                </div>
-                <div className="flex items-center justify-between">
-                  <div
-                    className="text-base text-white"
-                    style={{
-                      width: "40%",
-                      height: "20px",
-                      background: "gray",
-                    }}
-                  ></div>
-                  <div
-                    className="text-base text-white"
-                    style={{
-                      width: "40%",
-                      height: "20px",
-                      background: "gray",
-                    }}
-                  ></div>
-                </div>
-              </div>
-            </div>
-            <div
-              className="uppercase font-bold text-lg underline absolute bottom-[10px]"
-              style={{
-                width: "200px",
-                height: "20px",
-                background: "gray",
-              }}
-            ></div>
-          </div>
-          {/* bottom */}
-        </div>
-      ))}
-    </motion.div>
-  )
-}
